Improved the ExerciseForm
There is now a dropdown which contains a list of exercises completed before by using a GET request then mapping each added exercise objects name to a dropdown. This doesn't do anything functionally yet, it just is there visually.

TODO: Update implementation of exercise logging to shift to the new exercisesWithHistory format, which takes in a name, then an array with {BPM, time signature, and date} so that we can start to sort exercises by name.