Updated Exercise Submission Handling
The new implementation will use an input box to enter new exercises, which will be added to a database (the localhost 5001/exercises route).

The ExerciseSelector will be used to select from the existing exercises, then update this history of that exercise. This will ensure we are not creating duplicate exercises, and that we can consistently add to the history.

TODO: Implement exercise history updates upon submitting the "Update Exercise History"